Change to a simpler implementation based on Another-kev's solution
Now uses a) a regex to strip out any char that is not a lower case letter; b) a Set to capture the letters present in the given string (i.e. elements in Set are unique); and c) spread syntax to create the Set from the (iterable) stripped string.

The program then simply checks that the Set has 26 elements - i.e. all 26 distinct letters as elements.